Updating unit tests to stop patching in teardown - thru test_logger
